- Refactor POCOs
