+++ DarkUI/Controls/DarkDropdownList.cs
@@ -1,4 +1,6 @@
 ﻿using DarkUI.Config;
+using DarkUI.Data.Enums;
+using DarkUI.Data.Models;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
