+++ DarkUI/Data/Enums/Docking/DarkDockArea.cs
@@ -1,4 +1,4 @@
-﻿namespace DarkUI.Docking
+﻿namespace DarkUI.Data.Enums.Docking
 {
     public enum DarkDockArea
     {
