+++ DarkUI/Data/Models/DockGroupState.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace DarkUI.Docking
+namespace DarkUI.Data.Models
 {
     public class DockGroupState
     {
