+++ DarkUI/Docking/DarkDockContent.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DarkUI.Data.Enums.Docking;
+using System;
 using System.ComponentModel;
 using System.Drawing;
 using System.Windows.Forms;
