+++ DarkUI/Forms/DarkMessageBox.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DarkUI.Data.Enums;
+using System;
 using System.ComponentModel;
 using System.Drawing;
 using System.Windows.Forms;
