+++ Example/Forms/MainForm.cs
@@ -1,4 +1,5 @@
-﻿using DarkUI.Docking;
+﻿using DarkUI.Data.Models;
+using DarkUI.Docking;
 using DarkUI.Forms;
 using DarkUI.Win32;
 using System;
